django_project/requirements.txt
asgiref==3.2.6
Django==3.0.4
django-crispy-forms==1.9.0
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
